libraries/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pandaprobe"
dynamic = ["version"]
description = "Python SDK for PandaProbe — open-source agent tracing and evaluation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "PandaProbe | Chirpz AI", email = "support@chirpz.ai" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agents", "tracing", "observability", "evaluation", "llm"]

dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
gemini = ["google-genai>=1.0.0"]
anthropic = ["anthropic>=0.80.0"]
langgraph = ["langgraph>=0.2.0", "langchain-core>=0.3.0"]
all = ["pandaprobe[openai,gemini,anthropic,langgraph]"]
examples = ["pandaprobe[all]", "langchain-openai>=0.2.0"]
dev = [
    "pandaprobe[examples]",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "respx>=0.21.0",
    "ruff>=0.8.0",
    "build>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/chirpz-ai/pandaprobe-sdk"
Documentation = "https://github.com/chirpz-ai/pandaprobe-sdk"
Repository = "https://github.com/chirpz-ai/pandaprobe-sdk"

[tool.hatch.version]
path = "src/pandaprobe/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/pandaprobe"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 119
